# storage/api-docs/api-docs.json
{
    "swagger": "2.0",
    "info": {
        "title": "API Base Boilerplate",
        "description": "API Base for Laravel Boilerplate",
        "contact": {
            "email": "aldren.terante@gmail.com"
        },
        "version": "1.0.0"
    },
    "basePath": "/",
    "schemes": [
        "http"
    ],
    "paths": {
        "/api/hello": {
            "get": {
                "tags": [
                    "hello"
                ],
                "responses": {
                    "200": {
                        "description": "successful operation"
                    }
                }
            }
        },
        "/api/auth/login": {
            "post": {
                "tags": [
                    "auth"
                ],
                "summary": "Login.",
                "description": "Login.",
                "operationId": "login",
                "parameters": [
                    {
                        "name": "email",
                        "in": "query",
                        "description": "email",
                        "required": true,
                        "type": "string"
                    },
                    {
                        "name": "password",
                        "in": "query",
                        "description": "password",
                        "required": true,
                        "type": "string"
                    }
                ],
                "responses": {
                    "200": {
                        "description": ""
                    }
                }
            }
        },
        "/api/post": {
            "post": {
                "tags": [
                    "post"
                ],
                "summary": "Add a new post.",
                "description": "A user can add a new post.",
                "parameters": [
                    {
                        "name": "title",
                        "in": "query",
                        "description": "title",
                        "type": "string"
                    },
                    {
                        "name": "content",
                        "in": "query",
                        "description": "content",
                        "type": "string"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "this is a sample description"
                    }
                }
            }
        },
        "/api/post/{id}": {
            "patch": {
                "tags": [
                    "post"
                ],
                "summary": "update a post by id.",
                "description": "A user can update a post by id.",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "id",
                        "type": "string"
                    },
                    {
                        "name": "body",
                        "in": "body",
                        "description": "foo bar",
                        "required": true,
                        "schema": {}
                    }
                ],
                "responses": {
                    "200": {
                        "description": "this is a sample description"
                    }
                }
            }
        },
        "/api/auth/signup": {
            "post": {
                "tags": [
                    "auth"
                ],
                "summary": "Account signup for new users.",
                "description": "New account signup for new users.",
                "operationId": "signup",
                "parameters": [
                    {
                        "name": "name",
                        "in": "query",
                        "description": "name",
                        "required": true,
                        "type": "string"
                    },
                    {
                        "name": "email",
                        "in": "query",
                        "description": "email",
                        "required": true,
                        "type": "string"
                    },
                    {
                        "name": "password",
                        "in": "query",
                        "description": "password",
                        "required": true,
                        "type": "string"
                    }
                ],
                "responses": {
                    "200": {
                        "description": ""
                    }
                }
            }
        }
    },
    "definitions": {}
}